Refactor application structure: consolidate routes, remove unused pages, and implement MainRoutes for better navigation
